docs: add section for `tiebreakers` in migration guide (#77)
* docs(migrate): add section about `tiebreakers`

* docs: combine with upper heading level to form anchor link

* refactor: move mdx setup out from app file
