+++ src/docs/app.tsx
@@ -1,4 +1,4 @@
-import React, { forwardRef, isValidElement } from "react";
+import React from "react";
 import {
   BrowserRouter as Router,
   Navigate,
@@ -11,85 +11,17 @@ import preval from "preval.macro";
 
 import "./app.css";
 
-import { CodeBlock } from "./components/code-block";
-import linkIconSrc from "./assets/link.svg";
 import { DocsVersions } from "./views/docs-versions";
 import AppRoutes from "./app-routes";
 import Migrate from "./views/migrate.mdx";
 import "./utils/expose";
-
-function getAnchor(text: string) {
-  return text
-    .toLowerCase()
-    .replace(/[ \(\.]/g, "-")
-    .replace(/→/g, "to")
-    .replace(/[^a-z0-9-]/g, "");
-}
+import { mdxComponents } from "./utils/mdx-setup";
 
 const {
   fileVersions: docsVersions,
 }: {
   fileVersions: string[];
 } = preval`module.exports = require('./old-docs-list')`;
-
-interface HeadingProps extends React.HTMLAttributes<HTMLHeadingElement> {}
-
-const getHeading = (level: number) => {
-  const Heading = (props: HeadingProps, ref: React.Ref<HTMLHeadingElement>) => {
-    let anchor = getAnchor(
-      typeof props.children === "string" ? getAnchor(props.children) : ""
-    );
-    if (!anchor) {
-      if (isValidElement(props.children)) {
-        const elProps = props.children.props;
-        const children = elProps["children"];
-        if (typeof children === "string") {
-          anchor = getAnchor(children);
-        }
-      }
-    }
-
-    const link = `#${anchor}`;
-
-    return React.createElement(
-      `h${level}`,
-      {
-        id: anchor,
-        ref,
-      },
-      [
-        <a
-          key="1"
-          href={link}
-          className="heading-link"
-          style={{ textDecoration: "none" }}
-        >
-          {props.children}
-          <img
-            src={linkIconSrc}
-            className="w-5 inline-block ml-2"
-            style={{ marginTop: 0, marginBottom: 0 }}
-          />
-        </a>,
-      ]
-    );
-  };
-
-  return forwardRef(Heading);
-};
-
-const mdxComponents = {
-  wrapper: (props: any) => (
-    <div className="container mx-auto prose lg:max-w-3xl px-3 sm:px-0">
-      <main {...props} />
-    </div>
-  ),
-  code: CodeBlock as React.ComponentType<{ children: React.ReactNode }>,
-  // headings
-  ...Object.fromEntries(
-    [2, 3, 4].map((level) => ["h" + level, getHeading(level)])
-  ),
-};
 
 const oldDocs = docsVersions.map((version) => {
   return {
